requirements.txt
kafka-python==2.3.0
schedule==1.2.1